+++ database/migrations/2022_04_12_071101_create_additions_table.php
@@ -19,7 +19,7 @@ class CreateAdditionsTable extends Migration
             $table->unsignedBigInteger('employee_id');
             $table->unsignedBigInteger('payroll_id')->nullable();
             $table->date('date');  
-            $table->float('price', 10, 2)->default(0);
+            $table->decimal('price', 18, 8)->default(0);
             $table->string('description')->nullable();
             $table->string('type')->default('addition');
             $table->enum('status',['pending','paid','cancelled'])->default('pending');
