+++ database/migrations/2022_04_19_012228_create_income_tax_payroll_rules_table.php
@@ -16,9 +16,9 @@ class CreateIncomeTaxPayrollRulesTable extends Migration
         Schema::create('income_tax_payroll_rules', function (Blueprint $table) {
             $table->id();
             $table->foreignId('accounting_system_id')->constrained();
-            $table->float('income');
-            $table->float('rate');
-            $table->float('deduction');
+            $table->decimal('income', 18, 8);
+            $table->decimal('rate', 18, 8);
+            $table->decimal('deduction', 18, 8);
             $table->timestamps();
         });
     }
